Added better index page
